refactor: reorganize configuration settings by moving main settings to a new file
